+++ examples/table/demos/custom-col.vue
@@ -1,14 +1,16 @@
 <template>
   <div class="tdesign-demo-block-column-large">
     <!-- 按钮操作区域 -->
+    <!-- 设置哪些列允许自定义显示 :columnController="{ fields: ['platform', 'type', 'default']}" -->
     <t-table
       rowKey="index"
       :data="data"
       :columns="columns"
-      :stripe="stripe"
-      :bordered="bordered"
-      :hover="hover"
-      :columnController="{ displayType: 'auto-width' }"
+      :columnController="{ displayType: 'fixed-width', fields: ['platform', 'type', 'default'] }"
+      tableLayout="auto"
+      stripe
+      bordered
+      @column-change="onColumnChange"
     ></t-table>
   </div>
 </template>
@@ -21,7 +23,7 @@ for (let i = 0; i < 5; i++) {
     type: ['String', 'Number', 'Array', 'Object'][i % 4],
     default: ['-', '0', '[]', '{}'][i % 4],
     detail: {
-      postion: `读取 ${i} 个数据的嵌套信息值`,
+      position: `读取 ${i} 个数据的嵌套信息值`,
     },
     needed: i % 4 === 0 ? '是' : '否',
     description: '数据源',
@@ -31,19 +33,14 @@ export default {
   data() {
     return {
       data,
-      stripe: true,
-      bordered: true,
-      hover: false,
       columns: [
         {
           align: 'center',
-          width: '100',
           className: 'row',
           colKey: 'index',
           title: '序号',
         },
         {
-          width: 100,
           colKey: 'platform',
           title: '平台',
         },
@@ -60,39 +57,18 @@ export default {
           title: '是否必传',
         },
         {
-          colKey: 'detail.postion',
+          colKey: 'detail.position',
           title: '详情信息',
-          width: 200,
           ellipsis: true,
-          // 自定义 ellipsis 样式和内容，示例代码有效，勿删！！！
-          // ellipsis: (h, {
-          //   row, col, rowIndex, colIndex,
-          // }) => {
-          //   if (rowIndex % 2) {
-          //     return <div>is even row {rowIndex + 1}, with data {row.detail.postion}</div>;
-          //   }
-          //   return <div>is odd row {rowIndex + 1}, with data {row.detail.postion}</div>;
-          // },
         },
       ],
-      /** 非受控用法：与分页组件对齐 */
-      pagination: {
-        defaultCurrent: 2,
-        defaultPageSize: 10,
-        total: 120,
-      },
-      /** 受控用法：与分页组件对齐（此处注释为受控用法示例，代码有效，勿删） */
-      // pagination: {
-      //   current: 1,
-      //   pageSize: 10,
-      //   total: 120,
-      //   // 也可以监听表格组件的 page-change 事件进行处理
-      //   onChange: (pageInfo.current) => {
-      //     this.pagination.current = pageInfo.current;
-      //     this.pagination.pageSize = pageInfo.pageSize;
-      //   },
-      // },
     };
+  },
+
+  methods: {
+    onColumnChange(params) {
+      console.log(params);
+    },
   },
 };
 </script>