+++ examples/table/demos/filter-controlled.vue
@@ -4,65 +4,58 @@
       <t-button @click="setFilters">清除筛选条件</t-button>
       <span style="padding-left: 36px">已选筛选条件：{{ filterValue }}</span>
     </div>
+    <div>
+      <t-checkbox v-model="bordered">是否显示表格边框</t-checkbox>
+    </div>
 
-    <!-- 可以使用语法糖 .sync ，此处代码有效，勿删！ 其中，filterIcon 用于自定义筛选图标-->
+    <!-- 1. 此处代码有效，勿删！支持语法糖 filter-value.sync ， 支持非受控属性 defaultfilterValue -->
+    <!-- 2. 其中，filterIcon 用于自定义筛选图标，支持渲染函数 props.filterIcon，支持插槽 filterIcon。 -->
+    <!-- 3. filterRow={() => null}，则不会显示过滤行 -->
     <!-- <t-table
       rowKey='key'
       :columns="columns"
       :data="data"
       :filter-value.sync="filterValue"
       :filterIcon="filterIcon"
-    /> -->
+    >
+      <template #filterRow>自定义过滤行信息</template>
+    </t-table> -->
 
     <!-- filter-value.sync 等同于 filter-value + filter-change -->
-    <t-table rowKey="key" :columns="columns" :data="data" :filter-value="filterValue" @filter-change="onFilterChange" />
+    <!-- :filter-row="() => null" 用于隐藏过滤结果行 -->
+    <t-table
+      rowKey="key"
+      :columns="columns"
+      :data="data"
+      :filter-value="filterValue"
+      :bordered="bordered"
+      @filter-change="onFilterChange"
+      @change="onChange"
+    />
   </div>
 </template>
 
 <script lang="jsx">
-const data = [
-  {
-    key: '1',
-    firstName: 'Eric',
-    lastName: 'Spinke',
-    email: 'espinke0@apache.org',
-    createTime: '2021-11-01',
-  },
-  {
-    key: '2',
-    firstName: 'Gilberta',
-    lastName: 'Purves',
-    email: 'gpurves1@issuu.com',
-    createTime: '2021-12-01',
-  },
-  {
-    key: '3',
-    firstName: 'Heriberto',
-    lastName: 'Kment',
-    email: 'hkment2@nsw.gov.au',
-    createTime: '2022-01-01',
-  },
-  {
-    key: '4',
-    firstName: 'Lazarus',
-    lastName: 'Skures',
-    email: 'lskures3@apache.org',
-    createTime: '2022-02-01',
-  },
-  {
-    key: '5',
-    firstName: 'Zandra',
-    lastName: 'Croson',
-    email: 'zcroson5@virginia.edu',
-    createTime: '2022-03-01',
-  },
-];
+const data = new Array(5).fill(null).map((_, i) => ({
+  key: String(i + 1),
+  firstName: ['Eric', 'Gilberta', 'Heriberto', 'Lazarus', 'Zandra'][i % 4],
+  lastName: ['Spinke', 'Purves', 'Kment', 'Skures', 'Croson'][i % 4],
+  email: [
+    'espinke0@apache.org',
+    'gpurves1@issuu.com',
+    'hkment2@nsw.gov.au',
+    'lskures3@apache.org',
+    'zcroson5@virginia.edu',
+  ][i % 4],
+  createTime: ['2021-11-01', '2021-12-01', '2022-01-01', '2022-02-01', '2022-03-01'][i % 4],
+}));
 
 export default {
   data() {
     return {
       data,
       filterValue: {},
+      bordered: true,
       columns: [
         {
           title: 'FirstName',
@@ -81,6 +74,8 @@ export default {
         {
           title: 'LastName',
           colKey: 'lastName',
+          // 用于查看同时存在排序和过滤时的图标显示是否正常
+          sorter: true,
           // 多选过滤配置
           filter: {
             type: 'multiple',
@@ -111,7 +106,7 @@ export default {
         {
           title: 'Date',
           colKey: 'createTime',
-          // 自定义过滤组件
+          // 自定义过滤组件：日期过滤配置，请确保自定义组件包含 value 和 onChange 属性
           filter: {
             // eslint-disable-next-line @typescript-eslint/no-unused-vars
             component: (h) => <t-date-picker clearable />,
@@ -123,9 +118,14 @@ export default {
   methods: {
     // filters 参数包含自定义过滤组件 日期选择器 的值
     onFilterChange(filters) {
+      console.log('filter-change', filters);
       this.filterValue = filters;
       // 模拟异步请求进行数据过滤
       this.request(this.filterValue);
+    },
+    // 筛选、分页、排序等功能发生变化时，均会出发 change 事件
+    onChange(info, context) {
+      console.log('change', info, context, '筛选、分页、排序等功能发生变化均会触发');
     },
     setFilters() {
       this.filterValue = {};