+++ src/tabs/tab-panel.tsx
@@ -3,6 +3,7 @@ import props from './tab-panel-props';
 import { renderContent } from '../utils/render-tnode';
 import { prefix } from '../config';
 import Tabs from './tabs';
+import { updateElement } from '../hooks/useDestroyOnClose';
 
 export interface TabPanel extends Vue {
   parent?: InstanceType<typeof Tabs>;
@@ -26,6 +27,13 @@ export default (Vue as VueConstructor<TabPanel>).extend({
 
   updated() {
     this.parent?.updatePanels?.({ force: true });
+    // 父元素为 display: none 时，需要更新子元素，避免 Dialog 前套 Table 组件时，固定列等特性失效
+    if (!this.destroyOnHide) {
+      const timer = setTimeout(() => {
+        updateElement(this);
+        clearTimeout(timer);
+      }, 0);
+    }
   },
 
   render() {